requirements.txt
numpy==2.4.4
scipy==1.17.1
pandas==3.0.2
matplotlib==3.10.8
pyyaml==6.0.3
tqdm=4.67.3

torch==2.6.0+cu124

gymnasium==1.3.0
tensorboard==2.20.0

dimod==0.12.21
dwave-ocean-sdk==9.3.0
# dwave-neal
dwave-samplers==1.7.0

qiskit==2.4.0
qiskit-aer==0.17.2
qiskit-algorithms==0.4.0

pydantic==2.13.3
pytest==9.0.3